Repaired Privy.idr, made new tests in issue.idr
Privy.idr:
I had accidentaly unnoticely erased part of the file. Repaired it.

issue.idr:
The reason why it is a problem that an old fixity declaration is taken for
new shadowng operator definitions is that as a result of it you can not
declare a different fixity for your new operator :(
Not even by using a fully qualified name :/